feat : 1. 도메인 단위로 계층 분리 2. 테이블명을 제외하고 전부 단수형으로 바꾸기 3. 간단한 리팩토링
