+++ src/main/java/dbdr/domain/careworker/dto/request/CareworkerRequestDTO.java
@@ -1,0 +1,29 @@
+package dbdr.domain.careworker.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CareworkerRequestDTO {
+
+    @NotNull
+    private Long institutionId;
+
+    @NotBlank(message = "이름은 필수 항목입니다.")
+    private String name;
+
+    @NotBlank(message = "이메일은 필수 항목입니다.")
+    @Email(message = "올바르지 않은 형식입니다.")
+    private String email;
+
+    @NotBlank(message = "휴대폰 번호는 필수 항목입니다.")
+    @Pattern(regexp = "010\\d{8}", message = "010XXXXXXXX형식으로 입력해주세요.")
+    private String phone;
+}
