+++ src/main/java/dbdr/domain/chart/RecoveryTraining.java
@@ -1,6 +1,6 @@
 package dbdr.domain.chart;
 
-import dbdr.domain.BaseEntity;
+import dbdr.domain.core.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
