+++ src/main/java/dbdr/domain/guardian/repository/GuardianRepository.java
@@ -1,0 +1,9 @@
+package dbdr.domain.guardian.repository;
+
+import dbdr.domain.guardian.entity.Guardian;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GuardianRepository extends JpaRepository<Guardian, Long> {
+
+    boolean existsByPhone(String phone);
+}
