+++ src/main/java/dbdr/domain/recipient/dto/response/RecipientResponseDTO.java
@@ -1,0 +1,24 @@
+package dbdr.domain.recipient.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class RecipientResponseDTO {
+
+    private Long id;
+    private String name;
+    private LocalDate birth;
+    private String gender;
+    private String careLevel;
+    private String careNumber;
+    private LocalDate startDate;
+    private String institution;
+    private Long institutionNumber;
+    private Long careworkerId;
+}
