+++ src/main/java/dbdr/domain/recipient/repository/RecipientRepository.java
@@ -1,0 +1,7 @@
+package dbdr.domain.recipient.repository;
+
+import dbdr.domain.recipient.entity.Recipient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecipientRepository extends JpaRepository<Recipient, Long> {
+}
