--- src/main/java/dbdr/dto/response/CareworkerResponseDTO.java
@@ -1,17 +1,0 @@
-package dbdr.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CareworkerResponseDTO {
-
-    private Long id;
-    private Long institutionId;
-    private String name;
-    private String email;
-    private String phone;
-}
