--- src/main/java/dbdr/repository/CareworkerRepository.java
@@ -1,11 +1,0 @@
-package dbdr.repository;
-
-import dbdr.domain.Careworker;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface CareworkerRepository extends JpaRepository<Careworker, Long> {
-
-    List<Careworker> findByInstitutionId(Long institutionId);
-}
