--- src/main/java/dbdr/repository/GuardiansRepository.java
@@ -1,9 +1,0 @@
-package dbdr.repository;
-
-import dbdr.domain.Guardians;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GuardiansRepository extends JpaRepository<Guardians, Long> {
-
-    boolean existsByPhone(String phone);
-}
