--- src/main/java/dbdr/repository/RecipientRepository.java
@@ -1,8 +1,0 @@
-package dbdr.repository;
-
-import dbdr.domain.Recipient;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecipientRepository extends JpaRepository<Recipient, Long> {
-
-}
